不要なファイルやコードを削除
